daemon: Make sysroot_len be size_t instead of int. (cherry picked from commit 4b91462bad010a0b214e808f94a916002be65b91)

+++ daemon/daemon.h
@@ -36,7 +36,7 @@ extern int verbose;
 extern int autosync_umount;
 
 extern const char *sysroot;
-extern int sysroot_len;
+extern size_t sysroot_len;
 
 extern char *sysroot_path (const char *path);
